backend/pyzoz/requirements.txt
fastapi
uvicorn
python-socketio
pyserial
gpiozero
pydantic
sqlite3
